Set vel,ocity with nomalized values

+++ include/littlebot_firmware/motor_interface.h
@@ -29,7 +29,7 @@
 
 typedef struct MotorInterfaceClass MotorInterface;
 
-typedef void (*ptfSetVelocity)(MotorInterface *self, int16_t vel);
+typedef void (*ptfSetVelocity)(MotorInterface *self, float vel);
 typedef float (*ptfGetVelocity)(MotorInterface *self);
 typedef uint32_t (*ptfGetPosition)(MotorInterface *self);
 typedef uint8_t (*ptfGetSide)(MotorInterface *self);
@@ -49,7 +49,7 @@ struct MotorInterfaceClass{
 
 void MotorInterfaceConstruct(MotorInterface *self, uint8_t side);
 
-void fcSetVelocity(MotorInterface *self, int16_t velocity);
+void fcSetVelocity(MotorInterface *self, float velocity);
 
 float fcGetVelocity(MotorInterface *self);
